ikea-inbound/requirements.txt
requests==2.32.4
python-dotenv==1.1.1
paho-mqtt==2.1.0
websocket-client>=1.5.1